+++ database/migrations/2022_06_17_115201_create_purchase_details_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('purchase_details', function (Blueprint $table) {
             $table->id();
             $table->unsignedInteger('price');
-            $table->unsignedInteger('ppn');
+            $table->unsignedTinyInteger('ppn');
             $table->unsignedInteger('qty');
             $table->string('purchase_number');
             $table->string('product_number');
