+++ database/migrations/2022_06_27_105537_create_ppns_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('ppns', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('ppn');
+            $table->unsignedTinyInteger('ppn');
             $table->timestamps();
         });
     }
